+++ nextcloud-nginx-nomad/nextcloud-nginx-nomad.sh
@@ -15,10 +15,11 @@
 # 5. Adjust jail configuration script generation between BEGIN & END COOK
 #    Configure the config files that have been copied in where necessary
 
-# Set this to true if this jail flavour is to be created as a nomad (i.e. blocking) jail.
-# You can then query it in the cook script generation below and the script is installed
-# appropriately at the end of this script
-RUNS_IN_NOMAD=true
+# Set this to true if this jail flavour is to be created as a nomad
+# (i.e. blocking) jail.
+# You can then query it in the cook script generation below and the script
+# is installed appropriately at the end of this script
+RUNS_IN_NOMAD=false
 
 # set the cook log path/filename
 COOKLOG=/var/log/cook.log
@@ -57,12 +58,14 @@ set -e
 trap 'echo ERROR: $STEP$FAILED | (>&2 tee -a $COOKLOG)' EXIT
 
 # -------------- BEGIN PACKAGE SETUP -------------
+
 step "Bootstrap package repo"
 mkdir -p /usr/local/etc/pkg/repos
+# only modify repo if not already done in base image
 # shellcheck disable=SC2016
-#echo 'FreeBSD: { url: "pkg+http://pkg.FreeBSD.org/${ABI}/latest" }' \
-echo 'FreeBSD: { url: "pkg+http://pkg.FreeBSD.org/${ABI}/quarterly" }' \
-  >/usr/local/etc/pkg/repos/FreeBSD.conf
+test -e /usr/local/etc/pkg/repos/FreeBSD.conf || \
+  echo 'FreeBSD: { url: "pkg+http://pkg.FreeBSD.org/${ABI}/quarterly" }' \
+    >/usr/local/etc/pkg/repos/FreeBSD.conf
 ASSUME_ALWAYS_YES=yes pkg bootstrap
 
 step "Touch /etc/rc.conf"
@@ -75,7 +78,7 @@ step "Remove ifconfig_epair0b from config"
 sysrc -cq ifconfig_epair0b && sysrc -x ifconfig_epair0b || true
 
 step "Disable sendmail"
-service sendmail onedisable
+service sendmail onedisable || true
 
 # optionally disable ssh access
 step "Disable sshd"
@@ -213,16 +216,6 @@ pkg install -y php80-mbstring
 step "Install package php80-opcache"
 pkg install -y php80-opcache
 
-####
-# remove for php80, php82
-# doesn't exist, ssl may be integrated in core
-#step "Install package php74-openssl"
-#pkg install -y php74-openssl
-#
-#step "Install package php74-json"
-#pkg install -y php74-json
-####
-
 step "Install package php80-pcntl"
 pkg install -y php80-pcntl
 
@@ -295,6 +288,7 @@ pkg install -y sudo
 step "Install package syslog-ng"
 pkg install -y syslog-ng
 
+step "Clean package installation"
 pkg clean -y
 
 step "Create necessary directories if they don't exist"
@@ -305,14 +299,12 @@ step "Enable nginx"
 service nginx enable
 
 step "Enable php-fpm"
-#sysrc php_fpm_enable="YES"
 service php-fpm enable
 
-# ---------- END PACKAGE & MOUNTPOINT SETUP -------------
 
-#
-# Create configurations
-#
+
+
+# -------------- END PACKAGE SETUP -------------
 
 #
 # Now generate the run command script "cook"
@@ -320,216 +312,19 @@ service php-fpm enable
 # On subsequent runs, it only starts sleeps (if nomad-jail) or simply exits
 #
 
-# clear any old cook runtime file
-step "Clean cook artifacts"
-rm -rf /usr/local/bin/cook
-
+# this runs when image boots
 # ----------------- BEGIN COOK ------------------
-step "Create cook script"
-echo "#!/bin/sh
-RUNS_IN_NOMAD=$RUNS_IN_NOMAD
-# declare this again for the pot image, might work carrying variable through like
-# with above
-COOKLOG=/var/log/cook.log
 
-# No need to change this, just ensures configuration is done only once
-if [ -e /usr/local/etc/pot-is-seasoned ]
-then
-    # If this pot flavour is blocking (i.e. it should not return),
-    # we block indefinitely
-    if [ \"\$RUNS_IN_NOMAD\" = \"true\" ]
-    then
-        /bin/sh /etc/rc
-        tail -f /dev/null
-    fi
-    exit 0
-fi
+step "Clean cook artifacts"
+rm -rf /usr/local/bin/cook /usr/local/share/cook
 
-# ADJUST THIS: STOP SERVICES AS NEEDED BEFORE CONFIGURATION
-service nginx onestop || true
-service php-fpm onestop || true
+step "Install pot local"
+tar -C /root/.pot_local -cf - . | tar -C /usr/local -xf -
+rm -rf /root/.pot_local
 
-# No need to adjust this:
-# If this pot flavour is not blocking, we need to read the environment first from /tmp/environment.sh
-# where pot is storing it in this case
-if [ -e /tmp/environment.sh ]
-then
-    . /tmp/environment.sh
-fi
-
-#
-# ADJUST THIS BY CHECKING FOR ALL VARIABLES YOUR FLAVOUR NEEDS:
-#
-
-# Convert parameters to variables if passed (overwrite environment)
-while getopts d:s:r: option
-do
-    case \"\${option}\"
-    in
-      d) DATADIR=\${OPTARG};;
-      s) SELFSIGNHOST=\${OPTARG};;
-      r) REMOTELOG=\${OPTARG};;
-    esac
-done
-
-# Check config variables are set
-if [ -z \${DATADIR+x} ];
-then
-    echo 'DATADIR is unset - see documentation how to configure this flavour' >> /var/log/cook.log
-    echo 'DATADIR is unset - see documentation how to configure this flavour'
-    DATADIR=\"/usr/local/www/nextcloud/data\"
-fi
-if [ -z \${SELFSIGNHOST+x} ];
-then
-    echo 'SELFSIGNHOST is unset - see documentation how to configure this flavour' >> /var/log/cook.log
-    echo 'SELFSIGNHOST is unset - see documentation how to configure this flavour'
-    SELFSIGNHOST=\"none\"
-fi
-
-# ADJUST THIS BELOW: NOW ALL THE CONFIGURATION FILES NEED TO BE ADJUSTED & COPIED:
-
-# setup syslog-ng if remotelog var exists and isn't none
-# this needs some attention when moving to new image format, version 3.37 is hardcoded here
-if [ -n \"\${REMOTELOG}\" ]; then
-    if [ -f /root/syslog-ng.conf ]; then
-        < /root/syslog-ng.conf \
-           sed \"s|%%config_version%%|3.37|g\" | \
-           sed \"s|%%remotelogip%%|\$REMOTELOG|g\" \
-        > /usr/local/etc/syslog-ng.conf
-        service syslogd onestop || true
-        sysrc syslogd_enable=\"NO\"
-        # enable and start syslog-ng
-        service syslog-ng enable
-        sysrc syslog_ng_flags=\"-R /tmp/syslog-ng.persist\"
-        service syslog-ng start
-    fi
-fi
-
-# If we do not find a Nextcloud installation, we install it. If we do find something though,
-# we do not install/overwrite anything as we assume that updates/modifications are happening
-# from within the Nextcloud installation, we install it. If we do find something though,
-# we do not install/overwrite anything as we assume that updates/modifications are happening
-# from within Nextcloud.
-if [ ! -e /usr/local/www/nextcloud/status.php ]; then
-    pkg install -y nextcloud-php80 nextcloud-twofactor_totp-php80 nextcloud-deck-php80 nextcloud-mail-php80 nextcloud-contacts-php80 nextcloud-calendar-php80 nextcloud-end_to_end_encryption-php80
-fi
-
-# Configure PHP FPM
-sed -i .orig 's|listen = 127.0.0.1:9000|listen = /var/run/php80-fpm.sock|g' /usr/local/etc/php-fpm.d/www.conf
-sed -i .orig 's|pm.max_children = 5|pm.max_children = 10|g' /usr/local/etc/php-fpm.d/www.conf
-echo \";Nomad Nextcloud settings...\" >> /usr/local/etc/php-fpm.d/www.conf
-echo \"listen.owner = www\" >> /usr/local/etc/php-fpm.d/www.conf
-echo \"listen.group = www\" >> /usr/local/etc/php-fpm.d/www.conf
-echo \"listen.mode = 0660\" >> /usr/local/etc/php-fpm.d/www.conf
-echo \"env[PATH] = /usr/local/bin:/usr/bin:/bin\" >> /usr/local/etc/php-fpm.d/www.conf
-echo \"env[TMP] = /tmp\" >> /usr/local/etc/php-fpm.d/www.conf
-echo \"env[TMPDIR] = /tmp\" >> /usr/local/etc/php-fpm.d/www.conf
-echo \"env[TEMP] = /tmp\" >> /usr/local/etc/php-fpm.d/www.conf
-
-# Configure PHP
-cp -f /usr/local/etc/php.ini-production /usr/local/etc/php.ini
-cp -f /root/99-custom.ini /usr/local/etc/php/99-custom.ini
-
-# check for presence of copied-in /root/nc-config.php and copy over any existing (with backup)
-if [ -s /root/nc-config.php ]; then
-    if [ -s /usr/local/www/nextcloud/config/config.php ]; then
-        cp -f /usr/local/www/nextcloud/config/config.php /usr/local/www/nextcloud/config/config.php.old
-    fi
-    cp -f /root/nc-config.php /usr/local/www/nextcloud/config/config.php
-fi
-
-# Fix www group memberships so it works with fuse mounted directories
-pw addgroup -n newwww -g 1001
-pw moduser www -u 1001 -G 80,0,1001
-
-# set perms on /usr/local/www/nextcloud/*
-chown -R www:www /usr/local/www/nextcloud
-
-# create a nextcloud log file
-# this needs to be configured in nextcloud config.php in copy-in file
-touch /var/log/nginx/nextcloud.log
-chown www:www /var/log/nginx/nextcloud.log
-
-# manually create php log and set owner
-touch /var/log/nginx/php.scripts.log
-chown www:www /var/log/nginx/php.scripts.log
-
-# check for .ocdata in DATADIR
-# if using S3 with no mount-in this should set it up in the default DATADIR
-# /usr/local/nginx/nextcloud/data
-if [ ! -f \"\${DATADIR}\"/.ocdata ]; then
-   touch\"\${DATADIR}\"/.ocdata
-   chown www:www \"\${DATADIR}\"/.ocdata
-fi
-
-# set perms on DATADIR
-chown -R www:www \"\${DATADIR}\"
-
-# configure self-signed certificates for libcurl, mostly used for minio with self-signed certificates
-# nextcloud source needs patching to work with self-signed certificates too
-if [ \"\${SELFSIGNHOST}\" != \"none\" ]; then
-    echo \"\" | /usr/bin/openssl s_client -showcerts -connect \"\${SELFSIGNHOST}\" |/usr/bin/openssl x509 -outform PEM > /tmp/cert.pem
-    if [ -f /tmp/cert.pem ]; then
-        cat /tmp/cert.pem >> /usr/local/share/certs/ca-root-nss.crt
-        echo \"openssl.cafile=/usr/local/share/certs/ca-root-nss.crt\" >> /usr/local/etc/php/99-custom.ini
-        cat /tmp/cert.pem >> /usr/local/www/nextcloud/resources/config/ca-bundle.crt
-    fi
-    # Patch nextcloud source for self-signed certificates with S3
-    if [ -f /usr/local/www/nextcloud/lib/private/Files/ObjectStore/S3ObjectTrait.php ] && [ -f /root/S3ObjectTrait.patch ]; then
-        # make sure we haven't already applied the patch
-        checknotapplied=\$(grep -c verify_peer_name /usr/local/www/nextcloud/lib/private/Files/ObjectStore/S3ObjectTrait.php)
-        if [ \"\${checknotapplied}\" -eq 0 ]; then
-            # check the patch will apply cleanly
-            testpatch=\$(patch --check -i /root/S3ObjectTrait.patch /usr/local/www/nextcloud/lib/private/Files/ObjectStore/S3ObjectTrait.php | echo \"\$?\")
-            if [ \"\${testpatch}\" -eq 0 ]; then
-                # apply the patch
-                patch -i /root/S3ObjectTrait.patch /usr/local/www/nextcloud/lib/private/Files/ObjectStore/S3ObjectTrait.php
-            fi
-        fi
-    fi
-fi
-
-# Configure NGINX
-cp -f /root/nginx.conf /usr/local/etc/nginx/nginx.conf
-
-# ADJUST THIS: START THE SERVICES AGAIN AFTER CONFIGURATION
-
-# we need to kill nginx then start it
-killall -9 nginx || true
-# trying pkill instead
-pkill nginx || true
-#kill -9 \$(/bin/pgrep nginx)
-pkill php-fpm || true
-
-# restart services
-
-#start php-fpm
-timeout --foreground 120 \
-  sh -c 'while ! service php-fpm status; do
-    service php-fpm start || true; sleep 5;
-  done'
-
-#start nginx
-timeout --foreground 120 \
-  sh -c 'while ! service nginx status; do
-    service nginx start || true; sleep 5;
-  done'
-
-# setup cronjob
-echo \"*/15  *  *  *  *  www  /usr/local/bin/php -f /usr/local/www/nextcloud/cron.php\" >> /etc/crontab
-
-# Do not touch this:
-touch /usr/local/etc/pot-is-seasoned
-
-# If this pot flavour is blocking (i.e. it should not return), there is no /tmp/environment.sh
-# created by pot and we now after configuration block indefinitely
-if [ \"\$RUNS_IN_NOMAD\" = \"true\" ]
-then
-    /bin/sh /etc/rc
-    tail -f /dev/null
-fi
-" > /usr/local/bin/cook
-
+step "Set file ownership on cook scripts"
+chown -R root:wheel /usr/local/bin/cook /usr/local/share/cook
+chmod 755 /usr/local/share/cook/bin/*
 
 # ----------------- END COOK ------------------
 
@@ -591,7 +386,8 @@ then
   step "Enable cook service"
   # This is a non-nomad (non-blocking) jail, so we need to make sure the script
   # gets started when the jail is started:
-  # Otherwise, /usr/local/bin/cook will be set as start script by the pot flavour
+  # Otherwise, /usr/local/bin/cook will be set as start script by the pot
+  # flavour
   echo "enabling cook" | tee -a $COOKLOG
   service cook enable
 fi